requirements.txt
absl-py==2.0.0
asgiref==3.7.2
astunparse==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
flatbuffers==23.5.26
gast==0.5.4
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h5py==3.10.0
idna==3.4
importlib-metadata==6.8.0
keras==2.14.0
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
ml-dtypes==0.2.0
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
Pillow==10.0.1
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
werkzeug==3.0.0
wrapt==1.14.1
zipp==3.17.0
